Move operators to separate folder
