+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/__init__.py
@@ -1,7 +1,7 @@
 from .asset_condition import AssetCondition as AssetCondition
 
 # for whitelist_for_serdes
-from .dep_condition import (
+from .operators.dep_operators import (
     AllDepsCondition as AllDepsCondition,
     AnyDepsCondition as AnyDepsCondition,
 )
