+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/operands/slice_conditions.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 from dagster._core.asset_graph_view.asset_graph_view import AssetSlice
 
-from .asset_condition import AssetCondition, AssetConditionResult
-from .scheduling_context import SchedulingContext
+from ..asset_condition import AssetCondition, AssetConditionResult
+from ..scheduling_context import SchedulingContext
 
 
 class SliceSchedulingCondition(AssetCondition):
