+++ python_modules/dagster/dagster/_core/definitions/declarative_scheduling/operators/dep_operators.py
@@ -1,8 +1,8 @@
 from dagster._core.definitions.asset_key import AssetKey
 from dagster._serdes.serdes import whitelist_for_serdes
 
-from .asset_condition import AssetCondition, AssetConditionResult
-from .scheduling_context import SchedulingContext
+from ..asset_condition import AssetCondition, AssetConditionResult
+from ..scheduling_context import SchedulingContext
 
 
 class DepConditionWrapperCondition(AssetCondition):
